ICC fixes for throw on Exception-derived classes, and others
Summary:
1) #ifdef out __attribute__((cold)) for icc.
  2) Remove some trailing commas in enums to lower the noise level.
  3) Fix a cast of an uninitialized variable (this never executed,
     just cleans up a compiler warning).
  4) Move some forward declarations to be in front of the class
     definition (when trailing icc thinks it's a redeclare).
  5) 1<<31 is a negative number where a mask is wanted, fixed.
  6) "const CStrRef" is nonsense since it's passed by value.  fixed.
  7) Exception classes derived from Exception had destructors that
     implicitly permitted any throw, contrary to the ~Exception
     declaration.  Changed to declare "throw()" on all of them.
  8) "\{" is nonsense in a string.  Fixed to "\\{" as per hzhao.

Test Plan:
slow_tests, fast_tests

DiffCamp Revision: 146710
Reviewed By: hzhao
Commenters: mwilliams, macvicar
CC: hphp-diffs@lists, hzhao, macvicar, ervan, mwilliams, mpal
Revert Plan:
OK

+++ src/runtime/base/string_data.h
@@ -47,7 +47,7 @@ struct FilePlace {
  */
 class StringData {
  private:
-    const static unsigned int IsLiteral = (1 << 31); // literal string
+    const static unsigned int IsLiteral = ((unsigned)1 << 31); // literal string
     const static unsigned int IsShared  = (1 << 30); // shared memory string
     const static unsigned int IsLinear  = (1 << 29); // linear allocator memory
